Interpreter

+++ core/src/com/gordonfromblumberg/games/core/evocell/model/ActionMapping.java
@@ -4,5 +4,5 @@ import com.gordonfromblumberg.games.core.evocell.world.GameWorld;
 
 @FunctionalInterface
 public interface ActionMapping {
-    void act(GameWorld world, EvoLivingCell livingCell, byte parameter);
+    void act(GameWorld world, EvoLivingCell livingCell, int parameter);
 }
